Allows bulk upload in browsers that support HTML5
This essentially 2-line patch changes the form to use the [multiple
attribute] and uses the old view code to allow upload of more than one file
at once. This code still needs to be refactored and tested, so this is not a
definitive solution, but solves the biggest problem described by
namd/pypln.backend#114.

[multiple attribute]: http://www.w3.org/TR/html-markup/input.file.html#input.file.attrs.multiple