Add management command 'update_index'
This command search for documents that are not indexed and for each of them,
verify if the 'text' property exists. If it exists, it adds the document to the
index.
There is a weird exception raising at the end of running 'manage update_index'
related to GridFSStorage but as we are not using it in this command and all
operations are executed correctly, I'm ignoring it.