Creates Document slug automatically when saving it
This means we don't need to do it by hand in the view anymore.

Also, this commit introduces a test that depends on the mongo database, so we
clean it before and after the tests. To make we don't mess up our databases,
the TestCase checks for 'test' in the mongo database name before running the
tests. This commit also introduces a test settings file that should be used to
run tests.