Adds missing migrations
Some changes were made in the models in a branch different from the one that
introduced South in the project. After the merge, we didn't add migrations for
them. I tested this migration with a copy of the demo database in my localhost
and it seems to work fine, so we dodged a bullet here. We need to be more
careful about changes in the models.