Allows the creation of multiple pipelines at once
This means the manager is only created once, so we'll connect much fewer times
when more than one document is uplodaded.

Fixes #48 and helps with #43 (there seems to be some other bottleneck for that
one; in my development environment it takes a long time to process lots of
files before they hit the send pipelines step, but maybe in the server this
will be less noticeable).