Separates pos highlight from other visualizations (with minimal tests)
This is the only visualization that needs pagination, so it should be
different from the others. This commit also adds a minimal set of tests. The
idea is to simplify this view a lot, since it won't have to deal with the
different kinds of visualizations. For now, the view was only copied.