Adds basic pagination to the corpus document list
The javascript pagination does not work well for corpora that have a large
number of documents.

Next steps will be:

- allow the user to define the number of results per page
- allow sorting by filename, size, upload date and owner

We also need to implement pagination in the document list page (that is even
worse, since it shows all the documents a user has access to). In that case,
though, we can probably use [django's built-in object list
view](https://docs.djangoproject.com/en/1.3/ref/generic-views/#django-views-generic-list-detail-object-list)