Sets default page size to 100 documents
It used to take too long to load large corpora, specially when using the
browsable API, since the javascript processing necessary to pretty print the
json documents was too much for most browsers.

The default is 100, but the user can set the page size by sending 'page_size' as
a url parameter.