Uses the timestamp to create a unique name when inserting into GridFSStorage
Django will try to call `Storage.exists()`[1] multiple times, incrementing a
counter until it finds an available filename. This means that as the collection
grows with files uploaded using the same name (which is a normal usecase for
PyPLN) the insertion time also grows.

We are using a timestamp to include the new files with a (hopefully) unique
name. This gives chance for some a colision between two filenames, but we made
the decision of trading this chance for a faster insertion.

[1]: https://github.com/django/django/blob/master/django/core/files/storage.py#L90