Fixes a huge bug in the way `Storage.get_available_name` generated names
I mixed up old string formatting with new string formatting. The result is that
our GridFSStorage was always generating the same filename "%s_%s". It appears
gridfs does not care that much about the generated filename, and from a look at
the production database I could not see immediate effects, but the impact of
this bug is yet to be fully analysed. This is one to show when people ignore
the importance of tests and code reviews.