WIP: Updates base test to the new mongodb structure
We need to patch the `GridFSStorage.get_available_name()` here because our
files are being created in the setup, but the filenames are in the fixtures. If
we just left the method as it is, saving the file to gridfs would add another
timestamp to the filename and it would be unreachable.

This is a work in progress because it only creates the files for now, but we
still need to load the rest of the mongodb fixtures (the analyses themselves)
into the database.