Adds version specifiers in requirements that should include all minor versions
Following a discussion with @israelst[1] we decided that it would be a good
idea to keep our version specifiers. The main argument was that if some years
from now someone gets this code, they should be able to know which versions of
the software are going to work (and this probably would be the case if they
just download the newest version of these requirements then if pypln.web itself
isn't maintained anymore). When upgrading to new major versions we should
update the requirements file.

[1]: https://github.com/NAMD/pypln.web/pull/126#commitcomment-11504214