Makes sure the index name is prefixed with the user's username
The idea behind this is that a user should not be able to access or alter the
index for another user. By prefixing the index name with the username (and
enforcing this when the index is to be accessed) we make sure the user will
only access indexes created by himself.