Creates endpoint to return all document properties in one request
This is still a work in progess. There is an issue with UnicodeDecodeErrors
when returning all the keys. For some reason, when django rest framework gives
the returned dictionary to the renderer, every key is still a unicode object
(which is then passed to the json module, which cannot handle them). The weird
thing about this is that this does not happen in other places (the
"semantic_tags" property, for example, returns a dictionary with non-ascii
values and that is not a problem).