Adds new Storage class to store base64 encoded contents in MongoDB
We decided not to use GridFS. We're going to have a low (~3MB) file data size,
since the analysis can be as big as the file content itself. Since the workers
rely on having a 'contents' attribute in the analysis document anyways[0],
we'll just store it in MongoDB right away.

[0]: https://github.com/flavioamieiro/pypln.backend/blob/feature/remove_mongodict/pypln/backend/workers/extractor.py#L173