Fixes failures in create_pipeline tests
Mocking the create_pipeline_from_document object in its origin made some tests
fail when all tests were run. This happened because this object would remain as
a mock and later tests that expect this function to be actually called (and
expect to test if another function inside of it was called) would never reach
their goal.

Thanks @eduardostalinho for the report and fix.